Revert "Conditions 671 Update Feature Toggle for Platform ArrayField and Auto…"
This reverts commit 609c9ef72d56392ea64ae4ce0a3b327c17389664.
