+++ src/applications/disability-benefits/all-claims/components/NewDisability.jsx
@@ -2,12 +2,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { NULL_CONDITION_STRING } from '../constants';
+import { getShowAddDisabilitiesEnhancement } from '../content/addDisabilities';
 
 export default function NewDisability({ formData }) {
+  const showEnhancement = getShowAddDisabilitiesEnhancement();
+
   return (
-    // TODO https://github.com/department-of-veterans-affairs/vagov-claim-classification/issues/671:
-    // When remove allClaimsAddDisabilitiesEnhancement FF, remove the vads-u-flex--fill class
-    <div className="vads-u-margin-left--1 vads-u-flex--fill vads-u-padding-right--2 word-break capitalize-first-letter">
+    <div
+      className={`${
+        showEnhancement ? 'vads-u-margin-left--1' : 'vads-u-flex--fill'
+      } vads-u-padding-right--2 word-break capitalize-first-letter`}
+    >
       {typeof formData?.condition === 'string'
         ? formData.condition
         : NULL_CONDITION_STRING}
