+++ src/applications/disability-benefits/all-claims/content/addDisabilities.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router';
 
+import { SHOW_ADD_DISABILITIES_ENHANCEMENT } from '../constants';
+
 export const addDisabilitiesInstructions = (
   <>
     <h3>Tell us the new conditions you want to claim</h3>
@@ -72,3 +74,6 @@ export const increaseAndNewAlertRevised = ({ formContext }) => {
     </va-alert>
   );
 };
+
+export const getShowAddDisabilitiesEnhancement = () =>
+  window.sessionStorage.getItem(SHOW_ADD_DISABILITIES_ENHANCEMENT) === 'true';
