+++ src/applications/disability-benefits/all-claims/pages/addDisabilities.js
@@ -3,9 +3,12 @@ import set from 'platform/utilities/data/set';
 import get from 'platform/utilities/data/get';
 import omit from 'platform/utilities/data/omit';
 import fullSchema from 'vets-json-schema/dist/21-526EZ-ALLCLAIMS-schema.json';
+import * as combobox from '../definitions/combobox';
 import Autocomplete from '../components/Autocomplete';
 import disabilityLabelsRevised from '../content/disabilityLabelsRevised';
 import NewDisability from '../components/NewDisability';
+import ArrayField from '../components/ArrayField';
+import ConditionReviewField from '../components/ConditionReviewField';
 import {
   validateDisabilityName,
   requireDisability,
@@ -21,6 +24,7 @@ import {
 } from '../utils';
 import {
   addDisabilitiesInstructions,
+  getShowAddDisabilitiesEnhancement,
   increaseAndNewAlertRevised,
   newOnlyAlertRevised,
 } from '../content/addDisabilities';
@@ -48,6 +52,65 @@ const autocompleteUiSchema = {
   },
 };
 
+const comboboxUiSchema = combobox.uiSchema('Enter your condition', {
+  'ui:reviewField': ({ children }) => children,
+  'ui:options': {
+    debounceRate: 200,
+    freeInput: true,
+    inputTransformers: [
+      // Replace a bunch of things that aren't valid with valid equivalents
+      input => input.replace(/["”’]/g, `'`),
+      input => input.replace(/[;–]/g, ' -- '),
+      input => input.replace(/[&]/g, ' and '),
+      input => input.replace(/[\\]/g, '/'),
+      // TODO: Remove the period replacer once permanent fix in place
+      input => input.replace(/[.]/g, ' '),
+      // Strip out everything that's not valid and doesn't need to be replaced
+      // TODO: Add period back into allowed chars regex
+      input => input.replace(/([^a-zA-Z0-9\-',/() ]+)/g, ''),
+      // Get rid of extra whitespace characters
+      input => input.trim(),
+      input => input.replace(/\s{2,}/g, ' '),
+    ],
+    // options for the combobox dropdown
+    listItems: Object.values(disabilityLabelsRevised),
+  },
+  // autoSuggest schema doesn't have any default validations as long as { `freeInput: true` }
+  'ui:validations': [validateDisabilityName, limitNewDisabilities],
+  'ui:required': () => true,
+  'ui:errorMessages': {
+    required: missingConditionMessage,
+  },
+});
+
+const allClaimsArrayFieldWithCombobox = {
+  'ui:description': addDisabilitiesInstructions,
+  'ui:field': ArrayField,
+  'ui:options': {
+    viewField: NewDisability,
+    reviewTitle: 'Conditions',
+    duplicateKey: 'condition',
+    itemName: 'Condition',
+    itemAriaLabel: data => data.condition,
+    includeRequiredLabelInTitle: true,
+    classNames: 'cc-autocomplete-container',
+  },
+  useNewFocus: true,
+  // Ideally, this would show the validation on the array itself (or the name
+  // field in an array item), but that's not working.
+  'ui:validations': [requireDisability],
+  items: {
+    condition: comboboxUiSchema,
+    // custom review & submit layout - see https://github.com/department-of-veterans-affairs/vets-website/pull/14091
+    // disabled until design changes have been approved
+    'ui:objectViewField': ConditionReviewField,
+    'ui:options': {
+      itemAriaLabel: data => data.condition,
+      itemName: 'New condition',
+    },
+  },
+};
+
 const platformArrayFieldWithAutocomplete = {
   'ui:description': addDisabilitiesInstructions,
   'ui:options': {
@@ -69,7 +132,9 @@ const platformArrayFieldWithAutocomplete = {
 };
 
 export const uiSchema = {
-  newDisabilities: platformArrayFieldWithAutocomplete,
+  newDisabilities: getShowAddDisabilitiesEnhancement()
+    ? platformArrayFieldWithAutocomplete
+    : allClaimsArrayFieldWithCombobox,
   // This object only shows up when the user tries to continue without claiming either a rated or new condition
   'view:newDisabilityErrors': {
     'view:newOnlyAlert': {
