+++ src/applications/disability-benefits/all-claims/tests/fixtures/mocks/feature-toggles.json
@@ -17,10 +17,6 @@
       {
         "name": "subform_8940_4192",
         "value": true
-      },
-      {
-        "name": "all_claims_add_disabilities_enhancement",
-        "value": true
       }
     ]
   }
