+++ src/applications/disability-benefits/all-claims/tests/pages/addDisabilities.unit.spec.jsx
@@ -1,8 +1,7 @@
-import { fireEvent, render, waitFor } from '@testing-library/react';
+import { fireEvent, render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { expect } from 'chai';
 import { fullStringSimilaritySearch } from 'platform/forms-system/src/js/utilities/addDisabilitiesStringSearch';
-import { $, $$ } from 'platform/forms-system/src/js/utilities/ui';
 import { DefinitionTester } from 'platform/testing/unit/schemaform-utils';
 import set from 'platform/utilities/data/set';
 import React from 'react';
@@ -45,14 +44,6 @@ const createScreen = (
   );
 };
 
-const getVaButtonByText = (text, container) => {
-  return $(`va-button[text="${text}"]`, container);
-};
-
-const getAllVaButtonsByText = (text, container) => {
-  return $$(`va-button[text="${text}"]`, container);
-};
-
 const simulateInputChange = (selector, value) => {
   const vaTextInput = selector;
   vaTextInput.value = value;
@@ -64,54 +55,50 @@ const simulateInputChange = (selector, value) => {
   vaTextInput.dispatchEvent(event);
 };
 
-const addAConditionWithMouse = async (
-  getAllByRole,
-  getByTestId,
-  searchTerm,
-  searchResult,
-) => {
-  const input = getByTestId('autocomplete-input');
-  simulateInputChange(input, searchTerm);
-
-  await waitFor(() => {
-    const listResults = getAllByRole('option');
-
-    for (const result of listResults) {
-      if (result.textContent === searchResult) {
-        fireEvent.click(result);
-        const saveButton = getVaButtonByText('Save');
-        fireEvent.click(saveButton);
-      }
-    }
-  });
-};
-
-const addAConditionWithKeyboard = async (
+const addAConditionWithMouse = (
   getAllByRole,
   getByTestId,
   getByText,
   searchTerm,
   searchResult,
 ) => {
-  const input = getByTestId('autocomplete-input');
+  const input = getByTestId('combobox-input');
+  simulateInputChange(input, searchTerm);
+
+  const listboxItems = getAllByRole('option');
+  const freeTextItem = listboxItems.find(
+    item => item.textContent === searchResult,
+  );
+  fireEvent.click(freeTextItem);
+
+  const saveButton = getByText('Save');
+  fireEvent.click(saveButton);
+};
+
+const addAConditionWithKeyboard = (
+  getAllByRole,
+  getByTestId,
+  getByText,
+  searchTerm,
+  searchResult,
+) => {
+  const input = getByTestId('combobox-input');
   simulateInputChange(input, searchTerm);
 
   fireEvent.keyDown(input, { key: 'ArrowDown' });
 
-  await waitFor(() => {
-    const listResults = getAllByRole('option');
+  const listboxItems = getAllByRole('option');
 
-    for (const result of listResults) {
-      if (result.textContent !== searchResult) {
-        fireEvent.keyDown(result, { key: 'ArrowDown' });
-      } else if (result.textContent === searchResult) {
-        fireEvent.keyDown(input, { key: 'Enter' });
-        break;
-      }
+  for (const item of listboxItems) {
+    if (item.textContent !== searchResult) {
+      fireEvent.keyDown(item, { key: 'ArrowDown' });
+    } else if (item.textContent === searchResult) {
+      fireEvent.keyDown(input, { key: 'Enter' });
+      break;
     }
-  });
+  }
 
-  const saveButton = getVaButtonByText('Save');
+  const saveButton = getByText('Save');
   fireEvent.click(saveButton);
 };
 
@@ -171,14 +158,14 @@ describe('Add Disabilities Page', () => {
       expect(example7).to.be.visible;
     });
 
-    it('should render "Your new conditions" subheading, AutoComplete, save button, and add another condition button', () => {
-      const { container, getByRole, getByTestId, getByText } = createScreen();
+    it('should render "Your new conditions" subheading, ComboBox, save button, and add another condition button', () => {
+      const { getByRole, getByTestId, getByText } = createScreen();
 
       const newConditionsSubHeading = getByRole('heading', {
         name: 'Your new conditions',
       });
-      const input = getByTestId('autocomplete-input');
-      const saveButton = getVaButtonByText('Save', container);
+      const input = getByTestId('combobox-input');
+      const saveButton = getByText('Save');
       const addAnotherConditionButton = getByText('Add another condition');
 
       expect(newConditionsSubHeading).to.be.visible;
@@ -188,23 +175,29 @@ describe('Add Disabilities Page', () => {
     });
 
     it('should render with no saved conditions by default', () => {
-      const { container } = createScreen();
+      const { queryByText } = createScreen();
 
-      const savedConditionEditButton = getVaButtonByText('Edit', container);
+      const savedConditionEditButton = queryByText('Edit');
 
       expect(savedConditionEditButton).to.not.exist;
     });
 
-    it('should render autocomplete label with required and input', () => {
-      const { getByTestId } = createScreen();
+    it('should render ComboBox label with required, input, and listbox', () => {
+      const { getByRole, getByText, getByTestId } = createScreen();
 
-      const input = getByTestId('autocomplete-input');
+      const label = getByText('Enter your condition');
+      const required = label.querySelector('span').textContent;
+      const input = getByTestId('combobox-input');
+      const listbox = getByRole('listbox');
 
-      expect(input).to.have.attribute('label', 'Enter your condition');
-      expect(input).to.have.attribute('required');
+      expect(label).to.be.visible;
+      expect(required).to.eq('(*Required)');
+      expect(input).to.be.visible;
+      expect(listbox).to.be.visible;
+      expect(listbox).to.have.length(0);
     });
 
-    it('should render error message on result and alert on page if no new conditions are added', () => {
+    it('should render error message on item and alert on page if no new conditions are added', () => {
       const { getByText } = createScreen();
 
       const submitButton = getByText('Submit');
@@ -226,73 +219,63 @@ describe('Add Disabilities Page', () => {
 
   describe('Updating State', () => {
     it('should render with saved condition when there is initial formData', () => {
-      const { container, getByText } = createScreen(true, false, [
+      const { getByText } = createScreen(true, false, [
         {
           condition: 'asthma',
         },
       ]);
 
       const savedCondition = getByText('asthma');
-      const savedConditionEditButton = getVaButtonByText('Edit', container);
+      const savedConditionEditButton = getByText('Edit');
 
       expect(savedCondition).to.be.visible;
       expect(savedConditionEditButton).to.be.visible;
     });
 
-    it('should be able to add value to AutoComplete input ', async () => {
-      const searchTerm = 'a';
+    it('should be able to add value to ComboBox input ', () => {
+      const searchTerm = 'Typed value';
       const searchResults = fullStringSimilaritySearch(searchTerm, items);
       const freeTextAndFilteredItemsCount = searchResults.length + 1;
-      const { getByTestId, queryByTestId } = createScreen();
+      const { getByRole, getByTestId } = createScreen();
 
-      const input = getByTestId('autocomplete-input');
+      const input = getByTestId('combobox-input');
       simulateInputChange(input, searchTerm);
+      const listbox = getByRole('listbox');
 
-      await waitFor(() => {
-        const list = getByTestId('autocomplete-list');
+      expect(listbox).to.have.length(freeTextAndFilteredItemsCount);
 
-        expect(list).to.have.length(freeTextAndFilteredItemsCount);
-      });
+      fireEvent.click(document);
 
-      fireEvent.mouseDown(document.body);
-
-      await waitFor(() => {
-        const list = queryByTestId('autocomplete-list');
-
-        expect(list).to.not.exist;
-      });
+      expect(listbox).to.have.length(0);
+      expect(input).to.have.value(searchTerm);
     });
 
-    it('should render AutoComplete list items in alignment with string similarity search', async () => {
+    it('should render ComboBox listbox items in alignment with string similarity search', () => {
       const searchTerm = 'ACL';
       const searchResults = fullStringSimilaritySearch(searchTerm, items);
       const { getAllByRole, getByTestId } = createScreen();
 
-      const input = getByTestId('autocomplete-input');
+      const input = getByTestId('combobox-input');
       simulateInputChange(input, searchTerm);
+      const listboxItems = getAllByRole('option');
 
-      await waitFor(() => {
-        const listResults = getAllByRole('option');
-
-        listResults.forEach((result, index) => {
-          if (index === 0) {
-            expect(result.textContent).to.eq(
-              `Enter your condition as "${searchTerm}"`,
-            );
-          } else {
-            const searchResult = searchResults[index - 1];
-            expect(result.textContent).to.eq(searchResult);
-          }
-        });
+      listboxItems.forEach((item, index) => {
+        if (index === 0) {
+          expect(item.textContent).to.eq(
+            `Enter your condition as "${searchTerm}"`,
+          );
+        } else {
+          const searchResult = searchResults[index - 1];
+          expect(item.textContent).to.eq(searchResult);
+        }
       });
     });
   });
 
   describe('Mouse Interactions', () => {
-    it('should be able to add a free-text condition', async () => {
+    it('should be able to add a free-text condition', () => {
       const searchTerm = 'Tinnitus';
       const {
-        container,
         getAllByRole,
         getByTestId,
         getByText,
@@ -302,99 +285,84 @@ describe('Add Disabilities Page', () => {
       addAConditionWithMouse(
         getAllByRole,
         getByTestId,
+        getByText,
         searchTerm,
         `Enter your condition as "${searchTerm}"`,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton = getVaButtonByText('Edit', container);
-        const savedCondition = getByText(searchTerm);
+      const savedConditionEditButton = getByText('Edit');
+      const savedCondition = getByText(searchTerm);
 
-        expect(savedConditionEditButton).to.be.visible;
-        expect(savedCondition).to.be.visible;
+      expect(savedConditionEditButton).to.be.visible;
+      expect(savedCondition).to.be.visible;
 
-        const input = queryByTestId('autocomplete-input');
-        expect(input).to.not.exist;
-      });
+      const input = queryByTestId('combobox-input');
+      expect(input).to.not.exist;
     });
 
-    it('should be able to select a condition', async () => {
+    it('should be able to select a condition', () => {
       const searchTerm = 'Tinn';
       const searchResult = 'tinnitus (ringing or hissing in ears)';
-      const {
-        container,
-        getAllByRole,
-        getByTestId,
-        getByText,
-      } = createScreen();
+      const { getAllByRole, getByTestId, getByText } = createScreen();
 
       addAConditionWithMouse(
         getAllByRole,
         getByTestId,
+        getByText,
         searchTerm,
         searchResult,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton = getVaButtonByText('Edit', container);
-        const savedCondition = getByText(searchResult);
+      const savedConditionEditButton = getByText('Edit');
+      const savedCondition = getByText(searchResult);
 
-        expect(savedConditionEditButton).to.be.visible;
-        expect(savedCondition).to.be.visible;
-      });
+      expect(savedConditionEditButton).to.be.visible;
+      expect(savedCondition).to.be.visible;
     });
 
-    it('should be able to edit a condition', async () => {
+    it('should be able to edit a condition', () => {
       const searchTerm = 'Tinn';
       const searchResult = 'tinnitus (ringing or hissing in ears)';
       const newSearchTerm = 'Neck strain';
       const newSearchResult = 'neck strain (cervical strain)';
-      const {
-        container,
-        getAllByRole,
-        getByTestId,
-        getByText,
-      } = createScreen();
+      const { getAllByRole, getByTestId, getByText } = createScreen();
 
       addAConditionWithMouse(
         getAllByRole,
         getByTestId,
+        getByText,
         searchTerm,
         searchResult,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton = getVaButtonByText('Edit', container);
-        const savedCondition = getByText(searchResult);
+      const savedConditionEditButton = getByText('Edit');
+      const savedCondition = getByText(searchResult);
 
-        expect(savedConditionEditButton).to.be.visible;
-        expect(savedCondition).to.be.visible;
+      expect(savedConditionEditButton).to.be.visible;
+      expect(savedCondition).to.be.visible;
 
-        fireEvent.click(savedConditionEditButton);
-      });
+      fireEvent.click(savedConditionEditButton);
 
       addAConditionWithMouse(
         getAllByRole,
         getByTestId,
+        getByText,
         newSearchTerm,
         newSearchResult,
       );
+      const newCondition = getByText(newSearchResult);
 
-      await waitFor(() => {
-        const newCondition = getByText(newSearchResult);
-
-        expect(newCondition).to.be.visible;
-      });
+      expect(newCondition).to.be.visible;
     });
 
-    it('should be able to select two conditions then remove one', async () => {
+    it('should be able to select two conditions then remove one', () => {
       const searchTerm1 = 'Tinn';
       const searchResult1 = 'tinnitus (ringing or hissing in ears)';
       const searchTerm2 = 'Hear';
       const searchResult2 = 'hearing loss';
       const {
-        container,
         getAllByRole,
+        getAllByText,
         getByTestId,
         getByText,
         queryByText,
@@ -403,46 +371,40 @@ describe('Add Disabilities Page', () => {
       addAConditionWithMouse(
         getAllByRole,
         getByTestId,
+        getByText,
         searchTerm1,
         searchResult1,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton1 = getVaButtonByText('Edit', container);
-        const savedCondition1 = getByText(searchResult1);
+      const savedConditionEditButton1 = getByText('Edit');
+      const savedCondition1 = getByText(searchResult1);
 
-        expect(savedConditionEditButton1).to.be.visible;
-        expect(savedCondition1).to.be.visible;
-
-        const addAnotherConditionButton = getByText('Add another condition');
-        fireEvent.click(addAnotherConditionButton);
-      });
+      const addAnotherConditionButton = getByText('Add another condition');
+      fireEvent.click(addAnotherConditionButton);
 
       addAConditionWithMouse(
         getAllByRole,
         getByTestId,
+        getByText,
         searchTerm2,
         searchResult2,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton2 = getAllVaButtonsByText(
-          'Edit',
-          container,
-        )[1];
-        let savedCondition2 = getByText(searchResult2);
+      const savedConditionEditButton2 = getAllByText('Edit')[1];
+      let savedCondition2 = getByText(searchResult2);
 
-        expect(savedConditionEditButton2).to.be.visible;
-        expect(savedCondition2).to.be.visible;
+      expect(savedConditionEditButton1).to.be.visible;
+      expect(savedCondition1).to.be.visible;
+      expect(savedConditionEditButton2).to.be.visible;
+      expect(savedCondition2).to.be.visible;
 
-        fireEvent.click(savedConditionEditButton2);
-        const removeButton = getVaButtonByText('Remove', container);
-        fireEvent.click(removeButton);
+      fireEvent.click(savedConditionEditButton2);
+      const removeButton = getByText('Remove');
+      fireEvent.click(removeButton);
 
-        savedCondition2 = queryByText(searchResult2);
+      savedCondition2 = queryByText(searchResult2);
 
-        expect(savedCondition2).not.to.exist;
-      });
+      expect(savedCondition2).not.to.exist;
     });
 
     it('should submit when form is completed', () => {
@@ -470,10 +432,9 @@ describe('Add Disabilities Page', () => {
   });
 
   describe('Keyboard Interactions', () => {
-    it('should be able to add a free-text condition', async () => {
+    it('should be able to add a free-text condition', () => {
       const searchTerm = 'Tinnitus';
       const {
-        container,
         getAllByRole,
         getByTestId,
         getByText,
@@ -488,27 +449,20 @@ describe('Add Disabilities Page', () => {
         `Enter your condition as "${searchTerm}"`,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton = getVaButtonByText('Edit', container);
-        const savedCondition = getByText(searchTerm);
+      const savedConditionEditButton = getByText('Edit');
+      const savedCondition = getByText(searchTerm);
 
-        expect(savedConditionEditButton).to.be.visible;
-        expect(savedCondition).to.be.visible;
+      expect(savedConditionEditButton).to.be.visible;
+      expect(savedCondition).to.be.visible;
 
-        const input = queryByTestId('autocomplete-input');
-        expect(input).to.not.exist;
-      });
+      const input = queryByTestId('combobox-input');
+      expect(input).to.not.exist;
     });
 
-    it('should be able to select a condition', async () => {
+    it('should be able to select a condition', () => {
       const searchTerm = 'Tinn';
       const searchResult = 'tinnitus (ringing or hissing in ears)';
-      const {
-        container,
-        getAllByRole,
-        getByTestId,
-        getByText,
-      } = createScreen();
+      const { getAllByRole, getByTestId, getByText } = createScreen();
 
       addAConditionWithKeyboard(
         getAllByRole,
@@ -518,26 +472,19 @@ describe('Add Disabilities Page', () => {
         searchResult,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton = getVaButtonByText('Edit', container);
-        const savedCondition = getByText(searchResult);
+      const savedConditionEditButton = getByText('Edit');
+      const savedCondition = getByText(searchResult);
 
-        expect(savedConditionEditButton).to.be.visible;
-        expect(savedCondition).to.be.visible;
-      });
+      expect(savedConditionEditButton).to.be.visible;
+      expect(savedCondition).to.be.visible;
     });
 
-    it('should be able to edit a condition', async () => {
+    it('should be able to edit a condition', () => {
       const searchTerm = 'Tinn';
       const searchResult = 'tinnitus (ringing or hissing in ears)';
       const newSearchTerm = 'Neck strain';
       const newSearchResult = 'neck strain (cervical strain)';
-      const {
-        container,
-        getAllByRole,
-        getByTestId,
-        getByText,
-      } = createScreen();
+      const { getAllByRole, getByTestId, getByText } = createScreen();
 
       addAConditionWithKeyboard(
         getAllByRole,
@@ -547,15 +494,13 @@ describe('Add Disabilities Page', () => {
         searchResult,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton = getVaButtonByText('Edit', container);
-        const savedCondition = getByText(searchResult);
+      const savedConditionEditButton = getByText('Edit');
+      const savedCondition = getByText(searchResult);
 
-        expect(savedConditionEditButton).to.be.visible;
-        expect(savedCondition).to.be.visible;
+      expect(savedConditionEditButton).to.be.visible;
+      expect(savedCondition).to.be.visible;
 
-        userEvent.type(savedConditionEditButton, '{enter}');
-      });
+      userEvent.type(savedConditionEditButton, '{enter}');
 
       addAConditionWithKeyboard(
         getAllByRole,
@@ -564,22 +509,19 @@ describe('Add Disabilities Page', () => {
         newSearchTerm,
         newSearchResult,
       );
+      const newCondition = getByText(newSearchResult);
 
-      await waitFor(() => {
-        const newCondition = getByText(newSearchResult);
-
-        expect(newCondition).to.be.visible;
-      });
+      expect(newCondition).to.be.visible;
     });
 
-    it('should be able to select two conditions then remove one', async () => {
+    it('should be able to select two conditions then remove one', () => {
       const searchTerm1 = 'Tinn';
       const searchResult1 = 'tinnitus (ringing or hissing in ears)';
       const searchTerm2 = 'Hear';
       const searchResult2 = 'hearing loss';
       const {
-        container,
         getAllByRole,
+        getAllByText,
         getByTestId,
         getByText,
         queryByText,
@@ -593,16 +535,11 @@ describe('Add Disabilities Page', () => {
         searchResult1,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton1 = getVaButtonByText('Edit', container);
-        const savedCondition1 = getByText(searchResult1);
+      const savedConditionEditButton1 = getByText('Edit');
+      const savedCondition1 = getByText(searchResult1);
 
-        expect(savedConditionEditButton1).to.be.visible;
-        expect(savedCondition1).to.be.visible;
-
-        const addAnotherConditionButton = getByText('Add another condition');
-        userEvent.type(addAnotherConditionButton, '{enter}');
-      });
+      const addAnotherConditionButton = getByText('Add another condition');
+      userEvent.type(addAnotherConditionButton, '{enter}');
 
       addAConditionWithKeyboard(
         getAllByRole,
@@ -612,24 +549,21 @@ describe('Add Disabilities Page', () => {
         searchResult2,
       );
 
-      await waitFor(() => {
-        const savedConditionEditButton2 = getAllVaButtonsByText(
-          'Edit',
-          container,
-        )[1];
-        let savedCondition2 = getByText(searchResult2);
+      const savedConditionEditButton2 = getAllByText('Edit')[1];
+      let savedCondition2 = getByText(searchResult2);
 
-        expect(savedConditionEditButton2).to.be.visible;
-        expect(savedCondition2).to.be.visible;
+      expect(savedConditionEditButton1).to.be.visible;
+      expect(savedCondition1).to.be.visible;
+      expect(savedConditionEditButton2).to.be.visible;
+      expect(savedCondition2).to.be.visible;
 
-        userEvent.type(savedConditionEditButton2, '{enter}');
-        const removeButton = getVaButtonByText('Remove', container);
-        userEvent.type(removeButton, '{enter}');
+      userEvent.type(savedConditionEditButton2, '{enter}');
+      const removeButton = getByText('Remove');
+      userEvent.type(removeButton, '{enter}');
 
-        savedCondition2 = queryByText(searchResult2);
+      savedCondition2 = queryByText(searchResult2);
 
-        expect(savedCondition2).not.to.exist;
-      });
+      expect(savedCondition2).not.to.exist;
     });
 
     it('should submit when form is completed', () => {
@@ -657,10 +591,25 @@ describe('Add Disabilities Page', () => {
   });
 
   describe('Accessibility', () => {
+    it('should provide screen reader feedback when autocomplete results are available', () => {
+      const searchTerm = 'asthma';
+      const searchResults = fullStringSimilaritySearch(searchTerm, items);
+      const resultsCount = searchResults.length + 1;
+      const { getByTestId, getByText } = createScreen();
+
+      const input = getByTestId('combobox-input');
+      simulateInputChange(input, searchTerm);
+
+      const screenReaderMessage = getByText(
+        `${resultsCount} results available.`,
+      );
+      expect(screenReaderMessage).to.have.attribute('role', 'alert');
+    });
+
     it('should announce errors to screen readers when a required field is not filled', () => {
       const { getByTestId, getByText } = createScreen();
 
-      const input = getByTestId('autocomplete-input');
+      const input = getByTestId('combobox-input');
       const submitButton = getByText('Submit');
       simulateInputChange(input, '');
       fireEvent.click(submitButton);
