+++ src/applications/disability-benefits/all-claims/tests/property-names.unit.spec.js
@@ -60,10 +60,6 @@ describe('Root property names', () => {
       'view:hasEvidence',
       'view:selectableEvidenceTypes',
       'view:evidenceTypeHelp',
-      // TODO https://github.com/department-of-veterans-affairs/vagov-claim-classification/issues/671:
-      // When remove allClaimsAddDisabilitiesEnhancement FF, remove 'newDisabilities' and 'view:newDisabilityErrors as properties
-      'newDisabilities',
-      'view:newDisabilityErrors',
     ];
 
     Object.keys(pages).forEach(pageName => {
